requirements.txt
streamlit
ccxt
pandas
plotly
pyyaml
numpy
ta
matplotlib
seaborn
scipy
python-binance
pytest
python-dotenv
websocket-client
aiohttp
SQLAlchemy
pandas-ta
loguru
websockets
dash
dash-bootstrap-components